qcacmn: Enable WLAN driver build for canoe+wcn6450
Fix compilation errors for canoe+wcn6450

CRs-Fixed: 4369037

+++ hif/src/ce/ce_service.c
@@ -1081,7 +1081,7 @@ more_watermarks:
 					CE_state->id,
 					CE_state->dest_ring->nentries_mask,
 					CE_state->dest_ring->sw_index,
-					CE_DEST_RING_READ_IDX_GET(scn,
+					(uint32_t)CE_DEST_RING_READ_IDX_GET(scn,
 							  CE_state->ctrl_addr));
 			}
 		}
@@ -1103,7 +1103,7 @@ more_watermarks:
 					CE_state->src_ring->sw_index,
 					CE_state->src_ring->hw_index,
 					CE_state->src_ring->write_index,
-					CE_SRC_RING_READ_IDX_GET(scn,
+					(uint32_t)CE_SRC_RING_READ_IDX_GET(scn,
 							 CE_state->ctrl_addr));
 			}
 		}
